Added functionality for turning on/off model assumptions
Implemented the on/off functionality for the optional assumptions currently available in the framework implementation. The exact way of turning each one on/off in the code might not be how it should be in the long run, but at least the logic from a user perspective is ok I think.