Tweak of SCAgent.__init__() wrt optional assumptions
Previously if the argument optional_assumption was provided, it had to have a dict element for each OptionalAssumption in the current version of sc_scenario. This caused errors if loading a fit saved with an earlier version, after having added a new OptionalAssumption (in this case, oDA). So I have now changed it so that the SCAgent constructor first assumes that all optional assumptions are False, and then sets the ones provided by the caller (if any), to the values provided. (So now it is possible to call the constructor for example with just the few assumptions that one wants to enable.)